refactor!: remove deprecated GeneratedVaadinRadioButton
