--- vaadin-radio-button-flow-parent/vaadin-radio-button-flow/src/main/java/com/vaadin/flow/component/radiobutton/GeneratedVaadinRadioButton.java
@@ -1,285 +1,0 @@
-/*
- * Copyright 2000-2022 Vaadin Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-package com.vaadin.flow.component.radiobutton;
-
-import com.vaadin.flow.component.ClickNotifier;
-import com.vaadin.flow.component.Component;
-import com.vaadin.flow.component.ComponentEvent;
-import com.vaadin.flow.component.ComponentEventListener;
-import com.vaadin.flow.component.Focusable;
-import com.vaadin.flow.component.HasStyle;
-import com.vaadin.flow.component.Synchronize;
-import com.vaadin.flow.component.Tag;
-import com.vaadin.flow.component.dependency.JsModule;
-import com.vaadin.flow.component.dependency.NpmPackage;
-import com.vaadin.flow.shared.Registration;
-
-/**
- * <p>
- * Description copied from corresponding location in WebComponent:
- * </p>
- * <p>
- * {@code <vaadin-radio-button>} is a Web Component for radio buttons.
- * </p>
- * <p>
- * &lt;vaadin-radio-button
- * value=&quot;foo&quot;&gt;Foo&lt;/vaadin-radio-button&gt;
- * </p>
- * <h3>Styling</h3>
- * <p>
- * The following shadow DOM parts are available for styling:
- * </p>
- * <table>
- * <thead>
- * <tr>
- * <th>Part name</th>
- * <th>Description</th>
- * </tr>
- * </thead> <tbody>
- * <tr>
- * <td>{@code radio}</td>
- * <td>The radio button element</td>
- * </tr>
- * <tr>
- * <td>{@code label}</td>
- * <td>The label content element</td>
- * </tr>
- * </tbody>
- * </table>
- * <p>
- * The following state attributes are available for styling:
- * </p>
- * <table>
- * <thead>
- * <tr>
- * <th>Attribute</th>
- * <th>Description</th>
- * <th>Part name</th>
- * </tr>
- * </thead> <tbody>
- * <tr>
- * <td>{@code disabled}</td>
- * <td>Set when the radio button is disabled.</td>
- * <td>:host</td>
- * </tr>
- * <tr>
- * <td>{@code focus-ring}</td>
- * <td>Set when the radio button is focused using the keyboard.</td>
- * <td>:host</td>
- * </tr>
- * <tr>
- * <td>{@code focused}</td>
- * <td>Set when the radio button is focused.</td>
- * <td>:host</td>
- * </tr>
- * <tr>
- * <td>{@code checked}</td>
- * <td>Set when the radio button is checked.</td>
- * <td>:host</td>
- * </tr>
- * </tbody>
- * </table>
- * <p>
- * See
- * <a href="https://github.com/vaadin/vaadin-themable-mixin/wiki">ThemableMixin
- * – how to apply styles for shadow parts</a>
- * </p>
- */
-@Tag("vaadin-radio-button")
-@NpmPackage(value = "@vaadin/polymer-legacy-adapter", version = "24.0.0-alpha6")
-@JsModule("@vaadin/polymer-legacy-adapter/style-modules.js")
-@JsModule("@vaadin/radio-group/src/vaadin-radio-button.js")
-public abstract class GeneratedVaadinRadioButton<R extends GeneratedVaadinRadioButton<R>>
-        extends Component implements HasStyle, Focusable<R>, ClickNotifier<R> {
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * Specify that this control should have input focus when the page loads.
-     * <p>
-     * This property is not synchronized automatically from the client side, so
-     * the returned value may not be the same as in client side.
-     * </p>
-     *
-     * @return the {@code autofocus} property from the webcomponent
-     */
-    protected boolean isAutofocusBoolean() {
-        return getElement().getProperty("autofocus", false);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * Specify that this control should have input focus when the page loads.
-     * </p>
-     *
-     * @param autofocus
-     *            the boolean value to set
-     */
-    protected void setAutofocus(boolean autofocus) {
-        getElement().setProperty("autofocus", autofocus);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * If true, the user cannot interact with this element.
-     * <p>
-     * This property is not synchronized automatically from the client side, so
-     * the returned value may not be the same as in client side.
-     * </p>
-     *
-     * @return the {@code disabled} property from the webcomponent
-     */
-    protected boolean isDisabledBoolean() {
-        return getElement().getProperty("disabled", false);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * If true, the user cannot interact with this element.
-     * </p>
-     *
-     * @param disabled
-     *            the boolean value to set
-     */
-    protected void setDisabled(boolean disabled) {
-        getElement().setProperty("disabled", disabled);
-    }
-
-    /**
-     * This property is not synchronized automatically from the client side, so
-     * the returned value may not be the same as in client side.
-     *
-     * @return the {@code name} property from the webcomponent
-     */
-    protected String getNameString() {
-        return getElement().getProperty("name");
-    }
-
-    /**
-     * @param name
-     *            the String value to set
-     */
-    protected void setName(String name) {
-        getElement().setProperty("name", name == null ? "" : name);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * True if the radio button is checked.
-     * <p>
-     * This property is synchronized automatically from client side when a
-     * 'checked-changed' event happens.
-     * </p>
-     *
-     * @return the {@code checked} property from the webcomponent
-     */
-    @Synchronize(property = "checked", value = "checked-changed")
-    protected boolean isCheckedBoolean() {
-        return getElement().getProperty("checked", false);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * True if the radio button is checked.
-     * </p>
-     *
-     * @param checked
-     *            the boolean value to set
-     */
-    protected void setChecked(boolean checked) {
-        getElement().setProperty("checked", checked);
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * The value for this element.
-     * <p>
-     * This property is not synchronized automatically from the client side, so
-     * the returned value may not be the same as in client side.
-     * </p>
-     *
-     * @return the {@code value} property from the webcomponent
-     */
-    protected String getValueString() {
-        return getElement().getProperty("value");
-    }
-
-    /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
-     * The value for this element.
-     * </p>
-     *
-     * @param value
-     *            the String value to set
-     */
-    protected void setValue(String value) {
-        getElement().setProperty("value", value == null ? "" : value);
-    }
-
-    public static class CheckedChangeEvent<R extends GeneratedVaadinRadioButton<R>>
-            extends ComponentEvent<R> {
-        private final boolean checked;
-
-        public CheckedChangeEvent(R source, boolean fromClient) {
-            super(source, fromClient);
-            this.checked = source.isCheckedBoolean();
-        }
-
-        public boolean isChecked() {
-            return checked;
-        }
-    }
-
-    /**
-     * Adds a listener for {@code checked-changed} events fired by the
-     * webcomponent.
-     *
-     * @param listener
-     *            the listener
-     * @return a {@link Registration} for removing the event listener
-     */
-    protected Registration addCheckedChangeListener(
-            ComponentEventListener<CheckedChangeEvent<R>> listener) {
-        return getElement()
-                .addPropertyChangeListener("checked",
-                        event -> listener.onComponentEvent(
-                                new CheckedChangeEvent<R>((R) this,
-                                        event.isUserOriginated())));
-    }
-}
